Added delay after aborting test to allow the wic64 to time out properly
Previously if a test was aborted during a running transfer and the next
test was started immediately, the wic64 might not have been ready for the
next request because the previous transfer had not timed out properly yet.

Now added a slight delay after aborting a test to allow the wic64 to time
out properly (only where necessary).